Refactor sections into it's own component
